Add files via upload
This chart represents the average consumption of meat per capita per country, and highlights where this consumption stands relative to the guidelines of the French "Programme National Nutrition Santé" (2019-2023). The data is taken from the FAO (June 2023) as processed by Our World in Data. The relevant links are included in the chart.